backend/requirements.txt
Django>=4.2
djangorestframework
pymongo
python-dotenv
celery[redis]
playwright
gunicorn